Move TokenKindExtensions
